Added cascade type all to parkingSpot field of parkingZone, so parkingspots don't need to be persisted on their own. Removed setParkingSpots and added addParkingSpots, so that the parkingSpots will be more securely updated when a new parkingspot is added to the hashset.
